added ll support for knapsack. Plotted new knapsack results

+++ include/convolution.h
@@ -10,7 +10,7 @@ vector<int> boolCountConv(const vector<int>& a, const vector<int>& b);
 vector<long long> intSumConv(const vector<long long>& aSum, const vector<int>& b);
 
 // (max, +) convolution
-vector<int> maxPlusCnv(const vector<int>& a, const vector<int>& b);
+vector<ll> maxPlusCnv(const vector<ll>& a, const vector<ll>& b);
 
 // Boolean OR‐convolution (0/1 result)
 vector<int> boolCnv(const vector<int>& a, const vector<int>& b);
